Generating .ll and .lst for compiler-rt symbols
*) split into support and soft-float

+++ lib/patmos/getdecls.sh
@@ -4,17 +4,31 @@
 # and write declarations into a file.
 
 
-ARCHIVE=${1:-../../build/lib/libll.a}
+ARCHIVE=${1-../../build/lib/libllspt.a}
+if [ ! -f $ARCHIVE ]; then
+  echo File $ARCHIVE does not exist!
+  exit
+fi
+
+
+#LIBNAME=$(sed 's/lib\([[:alnum:]]*\)\.a/\1/' <(basename $ARCHIVE))
+OUT_BASE=${ARCHIVE%.a}syms
+OUT_LST=${OUT_BASE}.lst
+OUT_LL=${OUT_BASE}.ll
 
 echo Reading $ARCHIVE
-SYMBOLS=$(llvm-nm -defined-only $ARCHIVE | sed -n "s/^[\t ]*T \(.*\)$/@\1/p")
+llvm-nm -defined-only -extern-only $ARCHIVE | sed -n "s/^[\t ]*T \(.*\)$/\1/p"  > $OUT_LST
 
+# prepend each symbol with @
+SYMBOLS=$(sed 's/^.*$/@&/' < $OUT_LST)
+
+echo 'target triple = "patmos-unknown-elf"' > $OUT_LL
 llvm-ar x $ARCHIVE
 for f in *.bc
 do
   llvm-dis $f
   LLFILE=${f%.bc}.ll
   echo "; $LLFILE"
-  grep "$SYMBOLS" $LLFILE | grep 'define' | grep -v 'define internal' | sed -e "s/define/declare/" -e "s/ {$//"
+  grep "$SYMBOLS" $LLFILE | grep 'define' | sed -e "s/define/declare/" -e "s/ {$//"
   rm $f $LLFILE
-done > decls.ll
+done >> $OUT_LL
